Поправил баг в тесте

+++ src/test/groovy/com/example/benchmark/oracle/EditSpeedBenchmark.groovy
@@ -57,13 +57,13 @@ class EditSpeedBenchmark {
     void testChunks() {
 
         int i = 0
-        [20: 1, 100: 2, 500: 7, /*10_000: 134*/10_000: 137].each { fieldsCount, chunksCount ->
+        [20: 1, 100: 2, 500: 7, /*10_000: 134*/ 10_000: 137].each { fieldsCount, chunksCount ->
 
             List<Long> ids = readerService.loadChunkOperationIds(chunksCount)
             Assert.assertFalse("Operations not found in DB", ids.isEmpty())
 
             Map<Long, String> operations = ids.collectEntries { id ->
-                [id: readerService.readChunkOperation(id)]
+                [(id as Long): readerService.readChunkOperation(id)]
             }
 
             BenchmarkSuite.executeBenchmark(prepareReport(),
